requirements.txt
altair==5.5.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
attrs==25.1.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
blinker==1.9.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
cachetools==5.5.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
certifi==2024.12.14 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
charset-normalizer==3.4.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
click==8.1.8 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_full_version >= "3.11.4" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
contourpy==1.3.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
cycler==0.12.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
fastjsonschema==2.21.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
fonttools==4.55.6 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
gitdb==4.0.12 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
gitpython==3.1.44 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_full_version >= "3.11.4"
idna==3.10 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
isort==5.13.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
jinja2==3.1.5 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
joblib==1.4.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
jsonschema-specifications==2024.10.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
jsonschema==4.23.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
jupyter-core==5.7.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
kiwisolver==1.4.8 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
loguru==0.7.3 ; python_full_version >= "3.11.4" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
markupsafe==3.0.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
matplotlib==3.10.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
mdurl==0.1.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
narwhals==1.23.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
nbformat==5.10.4 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
numpy==2.2.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
packaging==24.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pandas==2.2.3 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pillow==11.1.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
platformdirs==4.3.6 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
plotly==5.24.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
protobuf==5.29.3 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
psycopg2==2.9.10 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pyarrow==19.0.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pydeck==0.9.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pygments==2.19.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pyparsing==3.2.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
python-dateutil==2.9.0.post0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
python-dotenv==1.0.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pytz==2024.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
pywin32==308 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_full_version >= "3.11.4" and python_full_version < "4.0.0"
referencing==0.36.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
requests==2.32.3 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
rich==13.9.4 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
rpds-py==0.22.3 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
six==1.17.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
smmap==5.0.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
sqlalchemy==2.0.37 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
streamlit==1.41.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
tenacity==9.0.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
toml==0.10.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
tornado==6.4.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
tqdm==4.67.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
traitlets==5.14.3 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
typing-extensions==4.12.2 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
tzdata==2025.1 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
urllib3==2.3.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
watchdog==6.0.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0" and platform_system != "Darwin"
win32-setctime==1.2.0 ; python_full_version >= "3.11.4" and python_version < "4.0" and sys_platform == "win32"
yapf==0.43.0 ; python_full_version >= "3.11.4" and python_full_version < "4.0.0"
